IR Received

+++ Firmware/Sample_Code/mikroC PRO for ARM/UARTprint.h
@@ -3,7 +3,7 @@
 
 void UARTDoublePrint(double value);
 void UARTIntegerPrint(int value);
-void UARTLongPrint(int value);
+void UARTLongPrint(long value);
 void UARTShortPrint(short value);
 void UARTLongHexPrint(unsigned long value);
